Build basic python package
