+++ eval_plus_secstr.py
@@ -7,16 +7,16 @@ import argparse
 
 import torch
 
-import src.config as config
-from src.data.alphabets import Protein
-import src.data.secstr as secstr
-import src.data.dataset as dataset
-import src.model.plus_rnn as plus_rnn
-import src.model.plus_tfm as plus_tfm
-import src.model.p_elmo as p_elmo
-import src.model.mlp as mlp
-from src.train import Trainer
-from src.utils import Print, set_seeds, set_output, load_models
+import plus.config as config
+from plus.data.alphabets import Protein
+import plus.data.secstr as secstr
+import plus.data.dataset as dataset
+import plus.model.plus_rnn as plus_rnn
+import plus.model.plus_tfm as plus_tfm
+import plus.model.p_elmo as p_elmo
+import plus.model.mlp as mlp
+from plus.train import Trainer
+from plus.utils import Print, set_seeds, set_output, load_models
 
 
 parser = argparse.ArgumentParser('Evaluate a Model on SecStr Datasets')
