+++ plus/config.py
@@ -13,7 +13,7 @@ from collections import OrderedDict
 
 import torch
 
-from src.utils import Print
+from plus.utils import Print
 
 
 class DataConfig():
