+++ plus/data/fasta.py
