+++ plus/model/plus_rnn.py
@@ -14,7 +14,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.utils.rnn import pack_padded_sequence, pad_packed_sequence
 
-from src.data.transmembrane import is_prediction_correct, Grammar
+from plus.data.transmembrane import is_prediction_correct, Grammar
 
 
 class PLUS_RNN(nn.Module):
