+++ pyproject.toml
@@ -1,0 +1,25 @@
+[tool.poetry]
+name = "PLUS"
+version = "0.1.0"
+description = "Protein sequence representations Learned Using Structural information (https://arxiv.org/abs/1912.05625)"
+authors = ["Seonwoo Min", "Seunghyun Park", "Siwon Kim", "Hyun-Soo Choi", "Sungroh Yoon"]
+
+[tool.poetry.urls]
+issues = "https://github.com/mswzeus/PLUS/issues"
+homepage = "http://ailab.snu.ac.kr/PLUS/"
+repository = "https://github.com/mswzeus/PLUS"
+
+[tool.poetry.dependencies]
+python = ">=3.6.1"
+torch = "^1.3.1"
+numpy = "^1.17.4"
+scipy = "^1.4.1"
+pandas = "^1.1.1"
+pillow = "^7.0.0"
+scikit-learn = "^0.22.1"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
